Add first version of EWC implementation.
Not sure if it fully works; there seems to be a "dip" in the accuracy
on the first dataset immediately after starting training on the second
one, but then the accuracy goes back up again. Could just be a matter
of choosing the Fisher coefficient?

For the Fisher information matrix, I used a sampling setup as
demonstrated here:
https://github.com/ariseff/overcoming-catastrophic